speed fix?

+++ app/src/main/java/de/jo/boattracker/util/SpeedConverter.java
@@ -2,11 +2,11 @@ package de.jo.boattracker.util;
 
 public class SpeedConverter {
 
-    public static double getMetersPerSecondToKnots(double ms) {
-        return ms * 1.943844;
+    public static float getMetersPerSecondToKnots(float ms) {
+        return ms * 1.943844f;
     }
 
-    public static double getKnotsToMetersPerSecond(double kn) {
-        return kn / 1.943844;
+    public static float getKnotsToMetersPerSecond(float kn) {
+        return kn / 1.943844f;
     }
 }
